#105: made KdTree abstract, with two concrete NTS and JTS implementation
