+++ NetTopologySuite/Index/KdTree/KdTree.cs
@@ -23,7 +23,7 @@ namespace NetTopologySuite.Index.KdTree
     /// <typeparam name="T">The type of the user data object</typeparam>
     /// <author>David Skea</author>
     /// <author>Martin Davis</author>
-    public class KdTree<T>
+    public abstract class KdTree<T>
         where T : class
     {
         ///<summary>
@@ -61,7 +61,7 @@ namespace NetTopologySuite.Index.KdTree
         }
 
 
-        private KdNode<T> _root;
+        protected KdNode<T> _root;
         private long _numberOfNodes;
         private readonly double _tolerance;
 
@@ -312,72 +312,11 @@ namespace NetTopologySuite.Index.KdTree
             QueryNode(_root, queryEnv, true, new KdNodeVisitor<T>(result));
         }
 
-        private static void NearestNeighbor(KdNode<T> currentNode,
-            Coordinate queryCoordinate, ref KdNode<T> closestNode, ref double closestDistanceSq,
-            bool isOddLevel)
-        {
-            while (true)
-            {
-                if (currentNode == null)
-                    return;
-
-
-                var distSq = Math.Pow(currentNode.X - queryCoordinate.X, 2) +
-                             Math.Pow(currentNode.Y - queryCoordinate.Y, 2);
-
-                if (distSq < closestDistanceSq)
-                {
-                    closestNode = currentNode;
-                    closestDistanceSq = distSq;
-                }
-
-                // Start with the branch that is more likely to produce the final result
-                var firstBranch = currentNode.Left;
-                var secondBranch = currentNode.Right;
-                if (currentNode.Left != null && (isOddLevel ? queryCoordinate.X >= currentNode.X : queryCoordinate.Y >= currentNode.Y))
-                {
-                    firstBranch = currentNode.Right;
-                    secondBranch = currentNode.Left;
-                }
-
-                if (firstBranch != null
-                    && NeedsToBeSearched(queryCoordinate, currentNode, closestDistanceSq, firstBranch == currentNode.Left, isOddLevel))
-                {
-                    NearestNeighbor(firstBranch, queryCoordinate, ref closestNode, ref closestDistanceSq, !isOddLevel);
-                }
-
-                if (secondBranch != null
-                    && NeedsToBeSearched(queryCoordinate, currentNode, closestDistanceSq, secondBranch == currentNode.Left, isOddLevel))
-                {
-                    currentNode = secondBranch;
-                    isOddLevel = !isOddLevel;
-                    continue;
-                }
-
-                break;
-            }
-        }
-
-        private static bool NeedsToBeSearched(Coordinate target,
-            KdNode<T> node, double closestDistSq,
-            bool left, bool isOddLevel)
-        {
-            if (isOddLevel)
-                return (left ? target.X <= node.X : target.X >= node.X) || Math.Pow(target.X - node.X, 2) < closestDistSq;
-            return (left ? target.Y <= node.Y : target.Y >= node.Y) || Math.Pow(target.Y - node.Y, 2) < closestDistSq;
-        }
-
         /// <summary>
         /// Performs a nearest neighbor search of the points in the index.
         /// </summary>
         /// <param name="coord">The point to search the nearset neighbor for</param>
-        public KdNode<T> NearestNeighbor(Coordinate coord)
-        {
-            KdNode<T> result = null;
-            var closestDistSq = double.MaxValue;
-            NearestNeighbor(_root, coord, ref result, ref closestDistSq, true);
-            return result;
-        }
+        public abstract KdNode<T> NearestNeighbor(Coordinate coord);
 
         private class KdNodeVisitor<T> : IKdNodeVisitor<T> where T : class
         {
