+++ PortableClassLibrary/NetTopologySuite.Pcl/NetTopologySuite.Pcl.csproj
@@ -94,6 +94,12 @@
     <Compile Include="..\..\NetTopologySuite\Index\KdTree\IKdNodeVisitor.cs">
       <Link>Index\KdTree\IKdNodeVisitor.cs</Link>
     </Compile>
+    <Compile Include="..\..\NetTopologySuite\Index\KdTree\KdTreeJTS.cs">
+      <Link>Index\KdTree\KdTreeJTS.cs</Link>
+    </Compile>
+    <Compile Include="..\..\NetTopologySuite\Index\KdTree\KdTreeNTS.cs">
+      <Link>Index\KdTree\KdTreeNTS.cs</Link>
+    </Compile>
     <Compile Include="..\..\NetTopologySuite\IO\WKTFileReader.cs">
       <Link>IO\WKTFileReader.cs</Link>
     </Compile>
